# cmd/docs/swagger.yaml
basePath: /v1
definitions:
...
  models.AnimeResponse:
    properties:
      created_at:
        description: Timestamp when the anime was created
        example: "2025-02-20T19:27:00Z"
        type: string
      description:
        description: Description of the anime
        example: A story about ninjas.
        type: string
      id:
        description: Unique identifier for the anime
        example: 1
        type: integer
      rating:
        description: Average rating of the anime
        example: 8.5
        type: number
      title:
        description: Title of the anime
        example: Naruto
        type: string
      updated_at:
        description: Timestamp when the anime was last updated
        example: "2025-02-20T19:27:00Z"
        type: string
    type: object
  models.Review:
    properties:
      animeId:
        description: ID of the anime being reviewed
        example: 1
        type: integer
      content:
        description: Content of the review
        example: This anime is amazing!
        maxLength: 500
        type: string
      created_at:
        description: Timestamp when the review was created
        example: "2025-02-20T19:27:00Z"
        type: string
      id:
        description: Unique identifier for the review
        example: 1
        type: integer
      rating:
        description: Rating given in the review (0-10)
        example: 9
        maximum: 10
        minimum: 0
        type: integer
      updated_at:
        description: Timestamp when the review was last updated
        example: "2025-02-20T19:27:00Z"
        type: string
      userId:
        description: ID of the user who created the review
        example: 1
        type: integer
    required:
    - content
    - rating
    type: object
  models.ReviewCreateRequest:
    properties:
      animeId:
        description: ID of the anime being reviewed
        example: 1
        type: integer
      content:
        description: Content of the review
        example: This anime is amazing!
        maxLength: 500
        type: string
      rating:
        description: Rating given in the review (0-10)
        example: 9
        maximum: 10
        minimum: 0
        type: integer
      userId:
        description: ID of the user creating the review
        example: 1
        type: integer
    required:
    - animeId
    - content
    - rating
    - userId
    type: object
  models.ReviewResponse:
    properties:
      anime:
        allOf:
        - $ref: '#/definitions/models.AnimeResponse'
        description: Anime details
      animeId:
        description: ID of the anime being reviewed
        example: 1
        type: integer
      content:
        description: Content of the review
        example: This anime is amazing!
        type: string
      created_at:
        description: Timestamp when the review was created
        example: "2025-02-20T19:27:00Z"
        type: string
      id:
        description: Unique identifier for the review
        example: 1
        type: integer
      rating:
        description: Rating given in the review (0-10)
        example: 9
        type: integer
      updated_at:
        description: Timestamp when the review was last updated
        example: "2025-02-20T19:27:00Z"
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.UserResponse'
        description: User details
      userId:
        description: ID of the user who created the review
        example: 1
        type: integer
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        description: Timestamp when the user was created
        example: "2025-02-20T19:27:00Z"
        type: string
      id:
        description: Unique identifier for the user
        example: 1
        type: integer
      is_admin:
        description: Indicates if the user has admin privileges
        example: false
        type: boolean
      updated_at:
        description: Timestamp when the user was last updated
        example: "2025-02-20T19:27:00Z"
        type: string
      username:
        description: Username of the user
        example: john_doe
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@myanimeapi.com
    name: API Support
  description: This is a sample API for managing anime and reviews.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: MyAnimeAPI
  version: "1.0"
paths:
...
  /v1/reviews:
    post:
      consumes:
      - application/json
      description: Create a new review for an anime
      parameters:
      - description: Review data
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.ReviewCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ReviewResponse'
        "400":
          description: Invalid input or missing required fields
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Anime or user not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create review
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new review
      tags:
      - reviews
  /v1/reviews/{id}:
    delete:
      description: Delete a review by its ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete review
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a review
      tags:
      - reviews
    get:
      description: Retrieve a review by its ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve review
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - "":
        - ""
      summary: Get a review by ID
      tags:
      - reviews
    put:
      consumes:
      - application/json
      description: Update an existing review with the provided data
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated review data
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.ReviewCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
        "400":
          description: Invalid input or ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update review
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a review
      tags:
      - reviews
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: 'Use the format "Bearer <JWT_TOKEN>". Example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"